+++ src/common/random-tasks.js
@@ -1,21 +1,30 @@
 function rollPercentageChance(percentVal)
 {
 	var randomSeed = Math.random();
-	var chanceRes = (randomSeed <= percentVal);
-	return chanceRes;
+	var outcomeRes = (randomSeed <= percentVal);
+	return outcomeRes;
 }
 
 
 function rollArrayElement(arrLength)
 {
 	var randomSeed = Math.random() * arrLength;
-	var choiceRes = Math.floor(randomSeed);
-	return choiceRes;
+	var outcomeRes = Math.floor(randomSeed);
+	return outcomeRes;
+}
+
+function rollNumberRange(lowerLimit, upperLimit)
+{
+	var numDiff = upperLimit - lowerLimit;
+	var randomSeed = Math.random() * numDiff;
+	var outcomeRes = Math.round(lowerLimit + randomSeed);
+	return outcomeRes;
 }
 
 
 module.exports =
 {
 	rollPercent: rollPercentageChance,
-	rollElement: rollArrayElement
+	rollElement: rollArrayElement,
+	rollNumber: rollNumberRange
 };